+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v2.9 [Unreleased]
 ### Features
+- [#726](https://github.com/hydra-billing/homs/pull/726/) Add services select table.
 - [#725](https://github.com/latera/homs/pull/725) Update Rails to 7
 - [#706](https://github.com/latera/homs/pull/706) Update Ruby to 3.2.2